requirements.txt
tencentcloud-sdk-python
nacos-sdk-python
pymongo
pymysql
django
djangorestframework
kubernetes
pipenv
django-cors-headers
cryptography
cos-python-sdk-v5
PyYaml